questivity-server/requirements.txt
asgiref==3.8.1
django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
mysqlclient==2.2.4
pyjwt==2.9.0
sqlparse==0.5.1
tzdata==2024.2
